fixes test for galleries

+++ app/models/gallery.rb
@@ -59,35 +59,39 @@ class Gallery < ActiveRecord::Base
   def published_outlets
     Outlet.where("draft_id IN (?)", self.outlet_ids)
   end
+  
   # This handles a json serialized format from the administrative end.
   # It is to allowed ordering of lists in the forms
   # Please be careful if messing with this, its senssitive
   def gallery_items_ol=(list)
-  	gallery_list = JSON.parse(list)[0]
-  	ids = []
-  	gallery_list.each_with_index do |item,index|
-  		if ["MobileApp","Outlet"].include? item["class"]
-  			if item["class"].constantize.find(item["id"])
-		  		new_item =  GalleryItem.find_or_create_by({
-		  			gallery_id: self.id,
-		  			item_id: item["id"],
-		  			item_type: item["class"]
-		  		})
-		  		new_item.item_order = index
-		  		new_item.save!
-		  		ids << item["id"]
-		  	else
-          # This error occurs if an invalid id is provided, generally should only be found by devs
-		  		self.errors.add(:base, "Couldn't find item to add to gallery")
-		  	end
-	  	else
-        # This error would require either a developer or something trying to do wrong to reach
-	  		self.errors.add(:base, "A gallery item was of the wrong class")
-	  	end
-  	end
-    # cleanup all records not found in the list this time. required due to way we are handling
-    # data serialization
-  	self.gallery_items.where('item_id NOT IN (?)', ids).destroy_all
+    if list
+    	gallery_list = JSON.parse(list)[0]
+    	ids = []
+    	gallery_list.each_with_index do |item,index|
+    		if ["MobileApp","Outlet"].include? item["class"]
+    			if item["class"].constantize.find(item["id"])
+  		  		new_item =  GalleryItem.find_or_create_by({
+  		  			gallery_id: self.id,
+  		  			item_id: item["id"],
+  		  			item_type: item["class"]
+  		  		})
+  		  		new_item.item_order = index
+  		  		new_item.save!
+  		  		ids << item["id"]
+  		  	else
+            # This error occurs if an invalid id is provided, generally should only be found by devs
+  		  		self.errors.add(:base, "Couldn't find item to add to gallery")
+  		  	end
+  	  	else
+          # This error would require either a developer or something trying to do wrong to reach
+  	  		self.errors.add(:base, "A gallery item was of the wrong class")
+  	  	end
+    	end
+      # cleanup all records not found in the list this time. required due to way we are handling
+      # data serialization
+    else
+    	self.gallery_items.where('item_id NOT IN (?)', ids).destroy_all
+    end
   end
